IT WORKS! Sort of.
Due to amateur design, I don't think it is truly possible to link up
InputThread and ChatActivity. At least, not right now. As it stands, the
thread gets the message from the irc server, parses & hands it to our
server object, who stashes it in the messages array. Over in
ChatActivity, it gets all the messages from said array, and sticks them
into the chatlog during onCreate(). Only on onCreate(). Maximum ghetto.